+++ src/lib/Operators/Import.ts
@@ -14,16 +14,28 @@
  *  limitations under the License.
  */
 
+import type { IDict } from '@litert/utils-ts-types';
 import * as cL from '../Constants';
 import type * as dL from '../Declaration';
 import * as _ from '../Utils';
+
+const FILES_TRAVERSED = Symbol('import:files_traversed');
 
 class ImportBlockOperator implements dL.IBlockOperator {
 
     public async process(operand: string, ctx: dL.IOperatorContext): Promise<void> {
 
-        const file = ctx.loader.reader.resolvePath(ctx.currentPath, operand);
-        const data = await ctx.loader.load(file, ctx.currentPath);
+        const filePath = ctx.loader.reader.resolvePath(ctx.currentPath, operand);
+
+        this._checkAndStart(ctx, filePath);
+
+        const data = await ctx.loader.load({
+            'path': filePath,
+            'parent': ctx.currentPath,
+            'contextData': ctx.data,
+        });
+
+        this._clean(ctx, filePath);
 
         if (_.isObject(ctx.output)) {
 
@@ -37,8 +49,17 @@ class ImportBlockOperator implements dL.IBlockOperator {
 
     public processSync(operand: string, ctx: dL.IOperatorContext): void {
 
-        const file = ctx.loader.reader.resolvePath(ctx.currentPath, operand);
-        const data = ctx.loader.loadSync(file, ctx.currentPath);
+        const filePath = ctx.loader.reader.resolvePath(ctx.currentPath, operand);
+
+        this._checkAndStart(ctx, filePath);
+
+        const data = ctx.loader.loadSync({
+            'path': filePath,
+            'parent': ctx.currentPath,
+            'contextData': ctx.data,
+        });
+
+        this._clean(ctx, filePath);
 
         if (_.isObject(ctx.output)) {
 
@@ -48,6 +69,25 @@ class ImportBlockOperator implements dL.IBlockOperator {
 
             ctx.output.push(data);
         }
+    }
+
+    private _checkAndStart(ctx: dL.IOperatorContext, file: string): void {
+
+        const filesTraversed = ctx.data[FILES_TRAVERSED] ??= {
+            [ctx.currentPath]: true,
+        } as IDict;
+
+        if (filesTraversed[file]) {
+
+            throw new Error(`Circular reference detected when importing from "${file}".`);
+        }
+
+        filesTraversed[file] = true;
+    }
+
+    private _clean(ctx: dL.IOperatorContext, file: string): void {
+
+        delete ctx.data[FILES_TRAVERSED]?.[file];
     }
 }